moved PrivateKeyConstructor to a test helper method and refactoring sk comparison
